# Resolve overhead of calling Haml::Buffer#format_script