I changed to generate `_hamlout.fix_textareas!(::Haml::Helpers.html_escape(i[:name].to_s))` instead of `_hamlout.format_script(i[:name],false,false,true,false)`.

With this change, we can avoid some unnecessary method calls' overhead without changing behaviors.

## benchmark
With Ruby 2.4.0 and [k0kubun/haml_bench/templates/slim_bench.haml](https://github.com/k0kubun/haml_bench/blob/833f04306bc138261d8d2102a4f0cceb3a78eca7/templates/slim_bench.haml),

### [before](https://travis-ci.org/k0kubun/haml_bench/jobs/205838755)

```
Rendering: /home/travis/build/k0kubun/haml_bench/templates/slim_bench.haml
Warming up --------------------------------------
          haml 4.0.7     1.648k i/100ms
   haml 5.0.0.beta.2     5.737k i/100ms
Calculating -------------------------------------
          haml 4.0.7     18.380k (± 2.4%) i/s -     92.288k in   5.024238s
   haml 5.0.0.beta.2     64.617k (± 1.4%) i/s -    327.009k in   5.061712s
Comparison:
   haml 5.0.0.beta.2:    64617.1 i/s
          haml 4.0.7:    18380.4 i/s - 3.52x  slower
```

### [after](https://travis-ci.org/k0kubun/haml_bench/jobs/205847380)
```
Rendering: /home/travis/build/k0kubun/haml_bench/templates/slim_bench.haml
Warming up --------------------------------------
          haml 4.0.7     1.441k i/100ms
   haml 5.0.0.beta.2     6.074k i/100ms
Calculating -------------------------------------
          haml 4.0.7     16.558k (± 8.0%) i/s -     83.578k in   5.084622s
   haml 5.0.0.beta.2     65.486k (±11.2%) i/s -    321.922k in   5.028385s
Comparison:
   haml 5.0.0.beta.2:    65486.2 i/s
          haml 4.0.7:    16558.4 i/s - 3.95x  slower
```